ignore custom automod and keeps warnings for admin+

+++ src/modules/AutoTriggerHandler.ts
@@ -250,6 +250,8 @@ export default class AutoTriggerHandler {
         const teamformingChannels = [this.client.channelIds.reminderChannel1, this.client.channelIds.reminderChannel2, this.client.channelIds.reminderChannel3, this.client.channelIds.reminderChannel4];
 
         if (teamformingChannels.includes(message.channelId) && (AutoTriggerHandler.keepsKeywords.some((keyword) => { return message.content.toLowerCase().includes(keyword)})) && 'send' in message.channel) {
+            if (await this.client.util.hasRolePermissionsMessage(this.client, ['admin', 'owner'], message)) return false;
+
             await message.channel.send(`<@${message.member?.id}> use <#1413114658541539410> for keeps!`);
             await message.delete();
             return true;
@@ -266,6 +268,8 @@ export default class AutoTriggerHandler {
     public async customAutomod(message: Message): Promise<boolean> {
         if (!message.inGuild()) return false;
         if (message.guildId !== process.env.GUILD_ID) return false;
+
+        if (await this.client.util.hasRolePermissionsMessage(this.client, ['admin', 'owner'], message)) return false;
 
         const adminChannelId = this.client.channelIds.ADMIN_CHANNEL;
         const adminChannel = await this.client.channels.fetch(adminChannelId) as TextChannel;
